i believe mif file has to be named
to match the actual memory variable used and the (* *) expression has to
come after the memory declaration.  changed the 32'd250000000 to
32d'50000000 so that the display changes once every second.  Let's ask
again in class about the issue you've raised with the values in the mif
file as well as the LED light madness.  If you're right, then it's best
to know b4 the assignment submission deadline.